pyproject.toml
[tool.poetry]
name = "hotel-tourist"
version = "0.1.0"
description = ""
authors = ["Nikita <Nikita.yavkin92@gmail.com>"]
readme = "README.md"
packages = [{include = "hotel_tourist"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
pillow = "^9.5.0"
listview = "^0.1.0"
django-user-accounts = "^3.2.0"
django-simple-captcha = "^0.5.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
